pyproject.toml
[tool.poetry]
name = "course-work-docker"
version = "0.1.0"
description = ""
authors = ["ssher <ssher250110@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
drf-spectacular = "^0.27.2"
redis = "^5.0.7"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
coverage = "^7.5.4"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
